arreglado el elegir tipoIncidencia

+++ incideitor-master/src/main/java/com/example/eoi/incideitor/repositories/IncidenciaRepository.java
@@ -3,5 +3,5 @@ package com.example.eoi.incideitor.repositories;
 import com.example.eoi.incideitor.entities.Incidencia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IncidenciaRepository extends JpaRepository<Incidencia, Integer> {
+public interface IncidenciaRepository extends JpaRepository<Incidencia, Long> {
 }
